+++ src/test/kotlin/br/com/study/codedesignpractice/book/controller/FindBookControllerTest.kt
@@ -80,7 +80,7 @@ class FindBookControllerTest {
         title = "Book One",
         summary = "First book summary",
         tableOfContents = "TOC 1",
-        price = 300,
+        price = 300.toBigDecimal(),
         numberOfPages = 200,
         isbn = "111-111-111",
         publishDate = LocalDate.now().plusDays(15),
