+++ src/test/kotlin/br/com/study/codedesignpractice/book/controller/response/BooksResponseTest.kt
@@ -51,7 +51,7 @@ class BooksResponseTest {
         title = "Book One",
         summary = "First book summary",
         tableOfContents = "TOC 1",
-        price = 300,
+        price = 300.toBigDecimal(),
         numberOfPages = 200,
         isbn = "111-111-111",
         publishDate = LocalDate.now().plusDays(15),
