+++ src/test/kotlin/br/com/study/codedesignpractice/purchase/PaymentControllerTest.kt
@@ -81,7 +81,7 @@ class PaymentControllerTest {
         title = "Book One",
         summary = "First book summary",
         tableOfContents = "TOC 1",
-        price = 300,
+        price = 300.toBigDecimal(),
         numberOfPages = 200,
         isbn = "111-111-111",
         publishDate = LocalDate.now().plusDays(15),
